Mover MCQPreviewPanel.jsx
