+++ src/components/MCQPanel/MCQPanel.jsx
@@ -2,43 +2,15 @@ import {
   Box,
   Container,
   MenuItem,
-  Paper,
   Select,
   TextField,
   Typography,
 } from "@mui/material";
 import * as React from "react";
 import { TabPanel } from "../TabPanel";
-import { BarChart } from "../BarChart";
-import { optionsToBarChartData } from "./utils";
 import { MCQOption } from "./MCQOption";
 import { AddOptionButton } from "./AddOptionButton";
-
-function MCQPreviewPanel({ question, options }) {
-  const paperSx = {
-    height: "95%",
-    width: "100%",
-    borderRadius: "0",
-    padding: "0.5em",
-  };
-  const containerSx = {
-    backgroundColor: "#E7E8EB",
-    display: "flex",
-    alignItems: "center",
-    width: "150%",
-  };
-
-  return (
-    <Container sx={containerSx}>
-      <Paper sx={paperSx}>
-        <Typography mb={4} variant="h4">
-          {question}
-        </Typography>
-        <BarChart data={optionsToBarChartData(options)} />
-      </Paper>
-    </Container>
-  );
-}
+import { MCQPreviewPanel } from "./MCQPreviewPanel";
 
 export function MCQPanel({ index, onChange, selectedTab, slide }) {
   function updateOption(optionIndex) {
